Move chroot command creation from Container.Chroot and Container.Run to global function chroot_command, to avoid unnecessary duplication
Add support for network namespaces, snd added Container.Port that maps a port in the namespace to a port in the host. Kill the socat used when done

When mounting /{dev,proc} for MacOS, instead of using bindfs, just mount the actual FS types

Instead of using "lsof -Fp", use "lsof -t"

When killing auxiliary processes, keep checking until all of them are gone. This is to prevent the situation where processes are started in the middle of a killing iteration

When calling Container.Chroot on a stopped container, start it, and wait until you can chroot. After, stop the container.